+++ src/components/NoteBox/index.js
@@ -1,5 +1,4 @@
 import React from 'react';
-import MDXRenderer from 'gatsby-plugin-mdx/mdx-renderer';
 import { Icon } from 'components';
 import './styles.scss';
 
@@ -10,8 +9,7 @@ export default function NoteBox({ children, title }) {
         <Icon img="notes" size={36} className="note-box__icon" />
         <span>{title}</span>
       </h3>
-      {/* <MDXRenderer>{children}</MDXRenderer> */}
-      {children}
+      4{children}
     </div>
   );
 }
